moved fastaccessdialog-related stuff to seperate source files
