+++ src/LogbookDialog.h
@@ -1200,29 +1200,4 @@ public:
 };
 #endif
 
-//////////////////////////////////////////////////////////////////////////////
-/// Class FastAccessDialog
-///////////////////////////////////////////////////////////////////////////////
-class FastAccessDialog : public wxDialog
-{
-	private:
-		wxGridSizer* container;
-		LogbookDialog* logbookDialog;
-
-	protected:
-
-		// Virtual event handlers, overide them in your derived class
-		virtual void FastAccessDialogOnInitDialog( wxInitDialogEvent& event ) { event.Skip(); }
-		void buttonCallbackHub( wxCommandEvent& event );
-
-
-	public:
-
-		FastAccessDialog( wxWindow* parent, wxWindowID id = wxID_ANY, const wxString& title = wxEmptyString, const wxPoint& pos = wxDefaultPosition, const wxSize& size = wxDefaultSize, long style = wxDEFAULT_DIALOG_STYLE );
-		~FastAccessDialog();
-
-		void AddButton(const wxString& title, bool toggleButton, void (*callback)(wxAnyButton* button, LogbookDialog* logbookDialog));
-
-};
-
 #endif
